fix: products is a separate module
