+++ src/app/accessories/entities/accessory.entity.ts
@@ -1,12 +1,12 @@
 import { Entity, Property } from '@mikro-orm/core';
 
-import { ProductCategory } from 'shared/enums/productCategory.enum';
+import { ProductEntity } from 'app/products/entities/product.entity';
+import { ProductCategories } from 'app/products/enums/product-categories.enum';
 
-import { ProductEntity } from '../../../shared/entities/product.entity';
 import { AccessoryRepo } from '../repo/accessories.repo';
 
 @Entity({
-  discriminatorValue: ProductCategory.ACCESSORIES,
+  discriminatorValue: ProductCategories.ACCESSORIES,
   customRepository: () => AccessoryRepo,
 })
 export class AccessoryEntity extends ProductEntity {
