+++ src/app/accessories/repo/accessories.repo.ts
@@ -1,7 +1,7 @@
 import { EntityRepository } from '@mikro-orm/postgresql';
 import { BadRequestException, Injectable } from '@nestjs/common';
 
-import { ProductCategory } from 'shared/enums/productCategory.enum';
+import { ProductCategories } from 'app/products/enums/product-categories.enum';
 
 import { AccessoryDTO } from '../dto/accessory.dto';
 import { CreateAccessoryForm } from '../dto/create-accessory.form';
@@ -42,7 +42,7 @@ export class AccessoryRepo extends EntityRepository<AccessoryEntity> {
   async createAccessory(accessoryData: CreateAccessoryForm) {
     const accessory = this.em.create(AccessoryEntity, {
       ...accessoryData,
-      category: ProductCategory.ACCESSORIES,
+      category: ProductCategories.ACCESSORIES,
     });
     await this.getEntityManager().persistAndFlush(accessory);
