+++ src/app/beer/entities/beer.entity.ts
@@ -1,12 +1,12 @@
 import { Entity, Property } from '@mikro-orm/core';
 
-import { ProductCategory } from 'shared/enums/productCategory.enum';
+import { ProductEntity } from 'app/products/entities/product.entity';
+import { ProductCategories } from 'app/products/enums/product-categories.enum';
 
-import { ProductEntity } from '../../../shared/entities/product.entity';
 import { BeerRepo } from '../repo/beer.repo';
 
 @Entity({
-  discriminatorValue: ProductCategory.BEER,
+  discriminatorValue: ProductCategories.BEER,
   customRepository: () => BeerRepo,
 })
 export class BeerEntity extends ProductEntity {
