+++ src/app/cart/cart.module.ts
@@ -3,9 +3,8 @@ import { Module } from '@nestjs/common';
 
 import { CartProductEntity } from 'app/cart-product/entities/cart-product.entity';
 import { CartProductRepo } from 'app/cart-product/repo/cart-product.repo';
-
-import { ProductEntity } from 'shared/entities/product.entity';
-import { ProductRepo } from 'shared/repo/product.repo';
+import { ProductEntity } from 'app/products/entities/product.entity';
+import { ProductRepo } from 'app/products/repo/product.repo';
 
 import { CartController } from './cart.controller';
 import { CartService } from './cart.service';
