+++ src/app/products/enums/product-categories.enum.ts
@@ -1,4 +1,4 @@
-export enum ProductCategory {
+export enum ProductCategories {
   BEER = 'beer',
   SNACKS = 'snacks',
   ACCESSORIES = 'accessories',
