+++ src/app/products/enums/product-types.enum.ts
