+++ src/app/snacks/dto/create-snack.form.ts
@@ -7,7 +7,7 @@ import {
   validate,
 } from 'class-validator';
 
-import { ProductTypes } from 'shared/enums/productTypes.enum';
+import { ProductTypes } from 'app/products/enums/product-types.enum';
 
 import { SnackTypes } from '../enums/snack-types.enum';
