+++ src/app/snacks/entities/snack.entity.ts
@@ -1,12 +1,12 @@
 import { Entity, Property } from '@mikro-orm/core';
 
-import { ProductCategory } from 'shared/enums/productCategory.enum';
+import { ProductEntity } from 'app/products/entities/product.entity';
+import { ProductCategories } from 'app/products/enums/product-categories.enum';
 
-import { ProductEntity } from '../../../shared/entities/product.entity';
 import { SnacksRepo } from '../repo/snack.repo';
 
 @Entity({
-  discriminatorValue: ProductCategory.SNACKS,
+  discriminatorValue: ProductCategories.SNACKS,
   customRepository: () => SnacksRepo,
 })
 export class SnacksEntity extends ProductEntity {
